Add start stream & stop stream integration test
