+++ webrtc-android-framework/build.gradle
@@ -75,6 +75,7 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     testImplementation 'org.mockito:mockito-core:1.10.19'
     testImplementation project(path: ':webrtc-android-framework')
+    testImplementation 'org.apache.commons:commons-lang3:3.12.0'
     androidTestImplementation 'androidx.test.ext:junit:1.1.4'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
 }
