+++ webrtc-android-sample-app/src/main/res/layout/activity_main.xml
@@ -55,4 +55,17 @@
         android:layout_centerHorizontal="true"
         android:layout_margin="3dp"
         />
+
+        <TextView
+            android:id="@+id/broadcasting_text_view"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:padding="10dp"
+            android:text="Broadcasting"
+            android:textColor="@android:color/holo_red_dark"
+            android:textSize="11sp"
+            android:visibility="gone"
+            android:layout_alignParentTop="true"
+            android:layout_alignParentEnd="true"
+            android:textStyle="bold"/>
 </RelativeLayout>
